Completed shipFactory function and test, making a start on the gameboard logic
Veered a little away from pure TDD, needed to get my head around how was best to
approach the gameboard function. Resorted to tried and tested manual test. Populated the
gameboard array and then convered this to a 10x10 grid of divs, with data attributes
corresponding to their position. Will now try and convert this back to a function
via TDD.